refactor: merged -> group
