+++ src/components/benefits-group-preview.vue
